+++ src/blocks/addMemeToTags.ts
@@ -1,13 +1,15 @@
-import { LocaleAwarePullRequest, PipelineConfig } from '../types/types';
-const fs = require('fs');
+import fs from 'fs';
 import path from 'path';
+
+import { LocaleAwarePullRequest, PipelineConfig } from '../types';
+
 export async function addMemeToTags(pullRequest: LocaleAwarePullRequest, config: PipelineConfig) {
   pullRequest.locales.map((locale) => {
     pullRequest[locale].meme.tags.forEach((tagId) => {
       const jsonPath = path.join(config.dbpath, 'l10n', locale, 'tags', `${tagId}.json`);
       let json;
       if (fs.existsSync(jsonPath)) {
-        json = JSON.parse(fs.readFileSync(jsonPath));
+        json = JSON.parse(fs.readFileSync(jsonPath, { encoding: 'utf-8' }));
       } else {
         json = {
           memes: [],
