+++ src/blocks/index.ts
@@ -1,0 +1,7 @@
+export * from './addMemeToTags';
+export * from './commitChanges';
+export * from './createMeme';
+export * from './createTags';
+export * from './get-pullrequests';
+export * from './push-changes';
+export * from './remove-pullrequests';
