+++ src/blocks/remove-pullrequests.ts
@@ -1,5 +1,6 @@
-import { LocaleAwarePullRequest, PipelineConfig } from '../types/types';
 import rimraf from 'rimraf';
+
+import { LocaleAwarePullRequest, PipelineConfig } from '../types';
 
 export async function removePullrequests(pullRequest: LocaleAwarePullRequest, config: PipelineConfig) {
   rimraf.sync(config.pullrequestsDir);
