+++ src/pr-pipeline.ts
@@ -1,11 +1,13 @@
-import { addMemeToTags } from './blocks/addMemeToTags';
-import { commitChanges } from './blocks/commitChanges';
-import { createMeme } from './blocks/createMeme';
-import { createTags } from './blocks/createTags';
-import { getPullrequests } from './blocks/get-pullrequests';
-import { removePullrequests } from './blocks/remove-pullrequests';
-import { PipelineConfig, LocaleAwarePullRequest } from './types/types';
-import { pushChanges } from './blocks/push-changes';
+import {
+  addMemeToTags,
+  commitChanges,
+  createMeme,
+  createTags,
+  getPullrequests,
+  pushChanges,
+  removePullrequests,
+} from './blocks';
+import { LocaleAwarePullRequest, PipelineConfig } from './types';
 
 export class PullRequestsPipeline {
   private blocks: Array<Function>;
