+++ src/types/data-set.ts
@@ -1,0 +1,12 @@
+import { LocaleAwarePullRequest } from './locale-aware-pull-reques';
+import { Meme } from './meme';
+import { Tag } from './tag';
+import { WithLocale } from './with-locale';
+
+export interface DataSet {
+  PullRequests: {
+    [key: string]: LocaleAwarePullRequest;
+  };
+  Memes: WithLocale<Meme[]>;
+  Tags: WithLocale<Tag[]>;
+}
