+++ src/types/locale-aware-pull-reques.ts
@@ -1,0 +1,8 @@
+import { Locale } from './locale';
+import { PullRequest } from './pull-request';
+
+export type LocaleAwarePullRequest = {
+  [key in Locale]?: PullRequest;
+} & {
+  locales: Locale[];
+};
