+++ src/types/locale.ts
@@ -1,0 +1,1 @@
+export type Locale = 'en' | 'ru';
