+++ src/types/meme.ts
@@ -1,0 +1,6 @@
+export interface Meme {
+  id?: string;
+  title: string;
+  description: string;
+  tags?: (string)[] | null;
+}
