+++ src/types/pipeline-config.ts
@@ -1,0 +1,4 @@
+export interface PipelineConfig {
+  dbpath: string;
+  pullrequestsDir: string;
+}
