+++ src/types/pull-request.ts
@@ -1,0 +1,6 @@
+import { Meme } from './meme';
+
+export interface PullRequest {
+  meme: Meme;
+  tags: string[];
+}
