+++ src/types/tag.ts
@@ -1,0 +1,4 @@
+export interface Tag {
+  id?: string;
+  title: string;
+}
