+++ src/types/with-locale.ts
@@ -1,0 +1,5 @@
+import { Locale } from './locale';
+
+export type WithLocale<T> = {
+  [key in Locale]: T;
+};
